Add files via upload
The Diabetes SQL Project involves analyzing a dataset related to diabetes patients using SQL. The dataset includes various health metrics for each patient, such as glucose levels, blood pressure, BMI, and age, along with an indicator of whether or not the patient has diabetes. The primary goal of this project is to use SQL queries to extract meaningful insights and patterns from the data, which can aid in understanding diabetes risk factors and potentially improve diagnosis and treatment.